refactor(projects): organize Project feature files by feature folder
Moves Project-related files (controller, commands, queries, request models, view model) from CQRS/Controllers file-type folders into a single src/SheetMusic.Api/Projects feature folder, renaming the namespace to SheetMusic.Api.Projects (plural, to avoid colliding with the Project entity type name).
