+++ .vscode/settings.json
@@ -11,6 +11,7 @@
         "git fetch": true,
         "git checkout": true,
         "git pull": true,
-        "git remote": true
+        "git remote": true,
+        "Test-Path": true
     }
 }
