+++ src/SheetMusic.Api/Projects/ConnectSetToProject.cs
@@ -5,7 +5,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace SheetMusic.Api.CQRS.Command;
+namespace SheetMusic.Api.Projects;
 
 public class ConnectSetToProject(Guid projectId, Guid setId) : IRequest
 {
