+++ src/SheetMusic.Api/Projects/DisconnectSetFromProject.cs
@@ -6,7 +6,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace SheetMusic.Api.CQRS.Command;
+namespace SheetMusic.Api.Projects;
 
 public class DisconnectSetFromProject(Guid projectId, Guid setId) : IRequest
 {
